tests/wpt/web-platform-tests/orientation-event/t003-manual.html

<!DOCTYPE html>
<html>
  <head>
    <title>User agents must also provide an event handler IDL attribute [HTML5] named ondeviceorientation on the window object</title>
    <meta name=viewport content="width=device-width, maximum-scale=1.0">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="author" title="Mosquito FP7">
    <link rel="reviewer author" title="Zhiqiang Zhang" href="mailto:zhiqiang.zhang@intel.com"> <!-- 2013-10-12 -->
  </head>
  <body>
    <p>Rotate the device to run the tests.</p>
    <div id="log"></div>
    <script>
      var t1 = async_test("Provide an event handler IDL attribute [HTML5] named ondeviceorientation");
      var t2 = async_test("The type of this event handler must be 'DeviceOrientationEvent'");
      var run = false;
      window.ondeviceorientation = function(e) {
        if (!run) {
          run = true;
          t1.step(function() {
            assert_equals(e.type, "deviceorientation");
          });
          t1.done();
          t2.step(function() {
            assert_true(e instanceof "DeviceOrientationEvent");
          });
          t2.done();
        }
      };
    </script>
  </body>
</html>
